Started implementation of A* search
This commit is just a placeholder to commit current start of modiciations to implement A*
Note that is largely not done yet and is to serve as an marker for other signficant changes like the use
of a shared_ptr in the Node class and changes to the priority queue as such. Also changes the comprator
from using std::greater to a lambda function to dereference the pointers first before comparison

+++ src/suas_planning/src/Waypoint.cpp
@@ -3,7 +3,7 @@
 namespace suas_planning {
     
 Waypoint::Waypoint(double x, double y) :
-    x_((unsigned int) round(x)),
-    y_((unsigned int) round(y)) {}
+    x_(round(x)),
+    y_(round(y)) {}
 
 }
